refactor:#172 home util 이동
