+++ src/app/(home)/components/home-calendar-week-item.tsx
@@ -3,9 +3,10 @@ import CircleProgressBar from '@/components/commons/circle-progress-bar';
 import { Typography } from '@/components/ui/typography';
 import { useCalendar } from '@/app/(home)/contexts/calendar.context';
 import { cn } from '@/lib/utils';
-import { formatDateWithDash, isSameDate } from '@/lib/utils/date.util';
+import { formatDateWithDash } from '@/lib/utils/date.util';
 import { getPercentage } from '@/lib/utils/nutrition-calculator.util';
 import { useDashboard } from '@/app/(home)/contexts/dashboard.context';
+import { isSameDate } from '../utils/calendar.util';
 
 type HomeCalendarWeekItemProps = {
   week: Date[];
