+++ src/app/(home)/components/home-macronutrient-group.tsx
@@ -1,5 +1,5 @@
-import MacronutrientBox from '@/components/_home/home-macronutrient-box';
 import { Macronutrient, MacronutrientEnum, MacronutrientGoal } from '@/types/nutrition.type';
+import HomeMacronutrientBox from './home-macronutrient-box';
 
 type HomeMacronutrientGroupProps = {
   total: Macronutrient;
@@ -9,17 +9,17 @@ type HomeMacronutrientGroupProps = {
 const HomeMacroNutrientGroup = ({ total, goal }: HomeMacronutrientGroupProps) => {
   return (
     <div className="flex gap-3">
-      <MacronutrientBox
+      <HomeMacronutrientBox
         variety={MacronutrientEnum.CARBOHYDRATE}
         value={total.carbohydrate}
         goal={goal ? goal.dailyCarbohydrateGoal : 0}
       />
-      <MacronutrientBox
+      <HomeMacronutrientBox
         variety={MacronutrientEnum.PROTEIN}
         value={total.protein}
         goal={goal ? goal.dailyProteinGoal : 0}
       />
-      <MacronutrientBox variety={MacronutrientEnum.FAT} value={total.fat} goal={goal ? goal.dailyFatGoal : 0} />
+      <HomeMacronutrientBox variety={MacronutrientEnum.FAT} value={total.fat} goal={goal ? goal.dailyFatGoal : 0} />
     </div>
   );
 };
