+++ src/app/(home)/utils/nutrition-diff.util.ts
